backend/requirements.txt
# API framework
fastapi==0.111.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Settings & validation
pydantic==2.7.1
pydantic-settings==2.3.1

# Database (async MongoDB)
motor==3.6.0
pymongo==4.9.2

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP client (LLM APIs + weather + translation)
httpx==0.27.0

# ML inference
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2
tensorflow-cpu==2.16.1
pillow==10.3.0

# Utilities
pyyaml==6.0.1
requests==2.32.3
